requirements/web.txt
django
django-unfold
djangorestframework
django-cors-headers
pytelegrambotapi
psycopg2
pytz
requests
celery
redis
yookassa
pytelegrambotapi
apscheduler
